Add 'Recently Loaded Files' option to file menu
Change the Slicer application to have a file menu option that contains
a list of recently loaded files. This list length is controlled by a spinbox in
the Slicer general settings.  The file list is created on startup and written out
on shutdown. When selected the addDataDialog is triggered with the file chosen already
in a row.

STILL TODO:
1. Adding additional files to the addDataDialog imports multiple copies of the
triggered data but with named as the files chosen

2. Move clear routines to a separated menu option.